skills/init/infrastructure/requirements.txt
pyyaml
filelock
